Update mobile banner to match new design
