+++ apps/web/src/contexts/globalLayout/GlobalBanner/MobileBetaReleaseBanner.tsx
@@ -4,6 +4,7 @@ import appIcon from 'public/gallery-app-ios-icon.png';
 import { useCallback } from 'react';
 import styled from 'styled-components';
 
+import { GlobalBanner } from '~/components/core/GlobalBanner/GlobalBanner';
 import { HStack, VStack } from '~/components/core/Spacer/Stack';
 import { TitleXSBold } from '~/components/core/Text/Text';
 import { useTrack } from '~/shared/contexts/AnalyticsContext';
@@ -30,6 +31,15 @@ export default function MobileBetaReleaseBanner({ handleCTAClick }: Props) {
 
     handleCTAClick();
   }, [handleCTAClick, push, track]);
+
+  return (
+    <GlobalBanner
+      title="Gallery"
+      description="Mobile app beta now available"
+      onClose={handleCTAClick}
+      onClick={handleClick}
+    />
+  );
 
   return (
     <StyledContainer align="center">
